feat: use GH Actions for build and deploy to NPM

+++ tasks/hooks.ts
@@ -6,15 +6,19 @@ import {setConfig} from 'Builder/taskList'
 export default {
   beforeAll: [
       async () => {
-        const ls = await fs.readdir(`${process.cwd()}/npm`);
-        for (const path of ls) {
-            // filter repo elements
-            const ls = (await fs.readdir(`${process.cwd()}/npm/${path}`)).filter(el => ![".git", ".gitignore", ".npmrc"].includes(el));
+        fs.exists(`${process.cwd()}/npm`, async (e) => {
+          if (e) {
+              const ls = await fs.readdir(`${process.cwd()}/npm`);
+              for (const path of ls) {
+                  // filter repo elements
+                  const ls = (await fs.readdir(`${process.cwd()}/npm/${path}`)).filter(el => ![".git", ".gitignore", ".npmrc"].includes(el));
 
-            for (const element of ls) {
-                await fs.rm(`${process.cwd()}/npm/${path}/${element}`, {recursive: true});
-            }
-        }
+                  for (const element of ls) {
+                      await fs.rm(`${process.cwd()}/npm/${path}/${element}`, {recursive: true});
+                  }
+              }
+          }
+        });
 
       },
     // () => fs.emptyDir(`${process.cwd()}/npm`),
